Move iframe helper functions into IframeUtils class
