+++ lib/msal-core/src/utils/UrlUtils.ts
@@ -6,6 +6,7 @@ import { Constants, SSOTypes } from "./Constants";
 import { ServerRequestParameters } from "../ServerRequestParameters";
 import { ScopeSet } from "../ScopeSet";
 import { StringUtils } from "./StringUtils";
+import { CryptoUtils } from "./CryptoUtils";
 
 /**
  * @hidden
@@ -215,5 +216,20 @@ export class UrlUtils {
             return urlStringOrFragment.substring(hashIndex1 + 1);
         }
         return urlStringOrFragment;
+    }
+
+    static urlContainsHash(urlString: string): boolean {
+        const parameters = UrlUtils.deserializeHash(urlString);
+        return (
+            parameters.hasOwnProperty(Constants.errorDescription) ||
+            parameters.hasOwnProperty(Constants.error) ||
+            parameters.hasOwnProperty(Constants.accessToken) ||
+            parameters.hasOwnProperty(Constants.idToken)
+        );
+    }
+
+    static deserializeHash(urlFragment: string) {
+        const hash = UrlUtils.getHashFromUrl(urlFragment);
+        return CryptoUtils.deserialize(hash);
     }
 }
